requirements.txt
Django==1.9.6
Markdown==2.6.3
PyJWT==1.4.0
boto==2.40.0
defusedxml==0.4.1
django-cors-headers==2.1.0
dj-database-url==0.4.0
django-braces==1.8.1
django-extensions==1.6.7
django-filter==0.11.0
django-oauth-toolkit==0.10.0
django-password-reset==0.8.2
django-storages==1.4.1
djangorestframework==3.2.4
gunicorn==19.4.5
oauthlib==1.0.3
psycopg2==2.7
python-social-auth==0.2.12
python3-openid==3.0.10
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
whitenoise==3.2
twilio==5.7.0
django-jquery-ui==1.11.4.1
validate_email==1.3
Pillow==4.0.0
certbot-django==0.1.0
django-bootstrap3-datetimepicker-2==2.4.2
django-rest-auth==0.9.1
django-allauth==0.32.0